コンテキストメニューの実装をMainFormが持ってしまっていたのを修正。
